Extract spacer construct into SpacerPattern recognizer
Move the inline spacer detection out of HtmlTransformer::convertElement
into a dedicated SpacerPattern in HtmlToBlocks/Patterns/, implementing the
PatternRecognizerInterface (match(DOMElement, PatternContext)) the same way
NavigationPattern does. The recognizer is self-contained: it reimplements
the small DOM helpers it needs (attr, hasClass, childElementCount,
spacerHeightFromStyle) and uses PatternContext for presentationAttributes
and createBlock.

convertElement now routes the div/section spacer check through
$this->spacerPattern->match() and the redundant private methods
spacerBlockFromElement() and spacerHeightFromStyle() are removed. The
recognizer is invoked only at the original call site (not added to the
shared firstMatch registry) so dispatch order and behavior stay
byte-identical.

Full suite green: contract + parity (124 fixtures) + packaging.
